+++ app/src/androidTest/java/com/monkeyteam/chimpagne/InstrumentedTest.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.test.performTextInput
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
 import com.monkeyteam.chimpagne.model.location.Location
-import com.monkeyteam.chimpagne.model.location.LocationHelper
+import com.monkeyteam.chimpagne.ui.utilities.MapContainer
 import com.monkeyteam.chimpagne.ui.theme.AccountCreation
 import kotlinx.coroutines.runBlocking
 import org.junit.Assert.*
@@ -67,18 +67,18 @@ class AccountCreationUITest {
   }
 }
 
-class TestLocationHelper {
+class TestMap {
   @Test
   fun check_adding_markers_works() = runBlocking {
-    val locationHelper = LocationHelper()
+    val map = MapContainer()
     val locParis = Location("Paris", 48.8566, 2.3522)
     val locBerlin = Location("Berlin", 52.5200, 13.4050)
     val locMadrid = Location("Madrid", 40.4168, 3.7038)
 
-    locationHelper.addMarker(locParis)
-    locationHelper.addMarker(locBerlin)
-    locationHelper.addMarker(locMadrid)
+    map.addMarker(locParis)
+    map.addMarker(locBerlin)
+    map.addMarker(locMadrid)
 
-    assertEquals(listOf(locParis, locBerlin, locMadrid), locationHelper.markers.value)
+    assertEquals(listOf(locParis, locBerlin, locMadrid), map.markers.value)
   }
 }
